Fixed pause bug
