+++ Assets/UI scripts/PauseButtonManager.cs
@@ -5,21 +5,23 @@ using UnityEngine.SceneManagement;
 
 public class PauseButtonManager : MonoBehaviour
 {
+    public PauseManager pauseManager;
+
     public void clickResume()
     {
-        Time.timeScale = 1f;
+        pauseManager.clickResume();
     }
 
     public void mainMenuButton()
     {
         if (SceneManager.GetActiveScene().name == "Start")
         {
-            Time.timeScale = 1f;
+            pauseManager.clickResume();
         }
         else
         {
+            pauseManager.clickResume();
             SceneManager.LoadScene("Start");
-            Time.timeScale = 1f;
         }
     }
 }
